log missed and hitted keys
