+++ pkg/cache/redis_cache.go
@@ -12,6 +12,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/trace"
 
 	util_log "github.com/grafana/tempo/pkg/util/log"
 	"github.com/grafana/tempo/pkg/util/spanlogger"
@@ -66,11 +67,8 @@ func (c *RedisCache) Fetch(ctx context.Context, keys []string) (found []string, 
 	const method = "RedisCache.MGet"
 	var items [][]byte
 	// Run a tracked request, using c.requestDuration to monitor requests.
-	err := instr.CollectedRequest(ctx, method, c.requestDuration, redisStatusCode, func(ctx context.Context) error {
-		log, _ := spanlogger.New(ctx, method)
-		defer log.End()
-		log.SetAttributes(attribute.Int("keys requested", len(keys)))
-
+	err := measureRequest(ctx, method, c.requestDuration, redisStatusCode, func(ctx context.Context) error {
+		log := spanlogger.FromContext(ctx)
 		var err error
 		items, err = c.redis.MGet(ctx, keys)
 		if err != nil {
@@ -79,9 +77,7 @@ func (c *RedisCache) Fetch(ctx context.Context, keys []string) (found []string, 
 			level.Error(c.logger).Log("msg", "failed to get from redis", "name", c.name, "err", err)
 			return err
 		}
-
-		log.SetAttributes(attribute.Int("keys found", len(items)))
-
+		log.AddEvent("cache.keys.found", trace.WithAttributes(attribute.Int("keys", len(keys))))
 		return nil
 	})
 	if err != nil {
@@ -104,9 +100,8 @@ func (c *RedisCache) Fetch(ctx context.Context, keys []string) (found []string, 
 func (c *RedisCache) FetchKey(ctx context.Context, key string) (buf []byte, found bool) {
 	const method = "RedisCache.Get"
 	// Run a tracked request, using c.requestDuration to monitor requests.
-	err := instr.CollectedRequest(ctx, method, c.requestDuration, redisStatusCode, func(ctx context.Context) error {
-		log, _ := spanlogger.New(ctx, method)
-		defer log.End()
+	err := measureRequest(ctx, method, c.requestDuration, redisStatusCode, func(ctx context.Context) error {
+		log := spanlogger.FromContext(ctx)
 		var err error
 		buf, err = c.redis.Get(ctx, key)
 		if err != nil {
@@ -114,15 +109,14 @@ func (c *RedisCache) FetchKey(ctx context.Context, key string) (buf []byte, foun
 			log.Error(err)
 			if errors.Is(err, redis.Nil) {
 				level.Debug(c.logger).Log("msg", "failed to get key from redis", "name", c.name, "err", err, "key", key)
+				log.AddEvent("cache.key.missed", trace.WithAttributes(attribute.String("key", key)))
 			} else {
 				level.Error(c.logger).Log("msg", "error requesting key from redis", "name", c.name, "err", err, "key", key)
 			}
 
 			return err
 		}
-
-		log.SetAttributes(attribute.String("key found", key))
-
+		log.AddEvent("cache.key.found", trace.WithAttributes(attribute.String("key", key)))
 		return nil
 	})
 	if err != nil {
